requirements.txt
appdirs==1.4.3
argh==0.26.2
arrow==0.14.4
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
beautifulsoup4==4.5.3
bs4==0.0.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
constantly==15.1.0
cryptography==2.4.2
cssselect==1.0.3
Django==2.1.4
django-background-tasks==1.2.0
django-compat==1.0.15
django-filter==2.0.0
django-widget-tweaks==1.4.3
djangorestframework==3.9.0
et-xmlfile==1.0.1
fake-useragent==0.1.11
Flask==1.1.1
html5lib==1.0.1
hyperlink==18.0.0
idna==2.7
ImageHash==4.0
incremental==17.5.0
ipaddress==1.0.22
itsdangerous==1.1.0
jaconv==0.2.4
jdcal==1.4
Jinja2==2.10.1
livereload==2.6.0
lxml==4.3.0
MarkupSafe==1.1.0
mws==0.8.6
numpy==1.15.4
olefile==0.46
opencv-python==4.1.0.25
openpyxl==2.5.0
parse==1.11.1
parsel==1.5.1
pathtools==0.1.2
Pillow==4.0.0
port-for==0.3.1
psutil==5.4.8
pyasn1==0.4.5
pyasn1-modules==0.2.3
pycparser==2.19
PyDispatcher==2.0.5
pyee==5.0.0
PyHamcrest==1.9.0
PyMySQL==0.9.2
pyOpenSSL==18.0.0
pyppeteer==0.0.25
pyquery==1.4.0
PySocks==1.6.8
pytesseract==0.2.6
python-dateutil==2.8.0
pytz==2018.7
PyWavelets==1.0.3
PyYAML==3.13
queuelib==1.5.0
redis==3.2.1
requests==2.13.0
requests-html==0.9.0
requests-toolbelt==0.8.0
rq==1.2.2
rq-dashboard==0.5.2
scipy==1.3.1
Scrapy==1.3.3
scrapy-amazon-robot-middleware3==0.3.3
service-identity==18.1.0
six==1.12.0
sphinx-autobuild==0.7.1
tornado==5.1.1
tqdm==4.30.0
Twisted==18.9.0
urllib3==1.24.1
uWSGI==2.0.17.1
w3lib==1.19.0
watchdog==0.9.0
webencodings==0.5.1
websockets==7.0
Werkzeug==0.15.5
zope.interface==4.6.0
